Initial support for Rp2040 architecture (#2392)
Raspberry PI Pico, etc.

**Initial RP2040 implementation**

Basic system working with:

- timers
- task callbacks
- flash memory/filing systems
- UART driver implementation

See arch. readme for details.

**Support non-recursive submodule imports**

The pico-sdk contains TinyUSB which isn't used, and is massive (c. 1.7GB).
Components can block recursive initialisation with a `.no-recursive` file.

**HostTests**

Allow building/running without network
Add System module with commonly-used system functions
FP calculations in Clock module can be slightly inaccurate, so allow error of 2 ticks

**Compile fixes**

Fix some compiler type mismatch errors.
Now builds on ARM architecture so invariably some discrepancies in system types.
In particular, printf format strings are a pain!

**Build/Installation/CI**

Add installer and CI builds, add HostTests to all architectures
Move complex AWK expression into script file
The 'help' output is generated with AWK, moved that into separate script file makes it easier to read.

+++ samples/Basic_Storage/app/application.cpp
@@ -49,7 +49,7 @@ void listSpiffsPartitions()
 
 void printPart(Storage::Partition part)
 {
-	size_t bufSize = std::min(4096U, part.size());
+	size_t bufSize = std::min(uint32_t(4096), part.size());
 	char buf[bufSize];
 	OneShotFastUs timer;
 	if(!part.read(0, buf, bufSize)) {
